change to use size_t instead of int is some places

+++ libft/ft_memcpy.c
@@ -6,13 +6,13 @@
 /*   By: jeremi360 <jbiernac@student.42warsaw.pl    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/29 18:57:59 by jeremi360         #+#    #+#             */
-/*   Updated: 2024/02/29 18:58:35 by jeremi360        ###   ########.fr       */
+/*   Updated: 2024/03/01 14:10:01 by jeremi360        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memcpy(void *dst, const void *src, int n)
+void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	unsigned char	*str1;
 	unsigned char	*str2;
