+++ libft/ft_memset.c
@@ -6,13 +6,13 @@
 /*   By: jeremi360 <jbiernac@student.42warsaw.pl    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/29 18:44:35 by jeremi360         #+#    #+#             */
-/*   Updated: 2024/02/29 18:58:37 by jeremi360        ###   ########.fr       */
+/*   Updated: 2024/03/01 14:09:00 by jeremi360        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *b, int c, int len)
+void	*ft_memset(void *b, int c, size_t len)
 {
 	unsigned char	*str;
